requirements.txt
httpx
pillow
coco-lib
tqdm
torch
torchvision
open-clip-torch
pandas